# core/vm: reuse Keccak-256 hashes across EVM opcode executions