Our SHA3 opcode implementation creates a new Keccak256 hasher for every invocation. This is quite wasteful, since resetting one is a lot faster than recreating it. This PR creates a single hasher within the EVM interpreted shared by the entire execution of a transaction. This is safe as opcodes always execute on a single thread.

As we've done before, we can also take advantage of the internals of the Keccak256 hasher (`Read` method) to extract the result without allocating a new slice via `Sum`. And yet a further optimization wrt memory allocations is to not reallocate the outside result buffer for every opcode, rather have a shared one in the interpreted and use that.

```
benchmark             old ns/op     new ns/op     delta
BenchmarkOpSHA3-8     1539          851           -44.70%

benchmark             old allocs     new allocs     delta
BenchmarkOpSHA3-8     6              2              -66.67%

benchmark             old bytes     new bytes     delta
BenchmarkOpSHA3-8     1086          123           -88.67%
```